Move data array arithmatic operations to arithmetic.cpp
